requirements.txt
torch==0.3
numpy
pandas
attrs
